requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
requests==2.13.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
